Se centraliza excepciones manejando excepciones personalizadas para el dominio

+++ domain/model/src/main/java/co/com/crediya/model/user/User.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -22,6 +23,6 @@ public class User {
     String identityDocument;
     String phoneNumber;
     Integer roleId;
-    Double baseSalary;
+    BigDecimal baseSalary;
 
 }
